Improve generate_schema.py 8.3.0 code
After talking for a bit on the click discord, I was pointed to the `to_info_dict` function.

Since the functionality of `to_info_dict` was preserved between 8.2.0 and 8.3.0, using it makes the code much cleaner while having the same effect.

Theoretically, the entirety of the function could re-use `to_info_dict`, but there are two problems: No `is_eager` detection, and it's currently untyped besides `dict[str, Any]`, so I decided against it and just did the one improvement.