Fix bug where module doc strings would be treated as normal strings if preceeded by comments
This happened because the code didn't account for comments before module doc strings. Since there can be arbitrarily many, this PR pulls the check into a new function so a while loop could go through all the previous blocks until either the start of the file is found, or a non comment block is reached. A new test was also added.