ci: add label for running all builds on a pull request
We currently don't run all builds on a pull request and only run them on main. This changes it to run all builds on a pull request when the 'ci: build all' label is added.

This is currently implemented in a hotfix manner which will only work once a commit is pushed *after* the label is added.